// interfaces/PointsOfInterest.ts
// Generated by https://quicktype.io

export interface FetchedPointsOfInterest {
    list:  PointOfInterest[];
    _meta: Meta;
}

export interface Meta {
    isCached: null;
    count:    number;
}

export interface PointOfInterest {
    place_id:           string;
    photo_reference:    string;
    latitude:           number;
    longitude:          number;
    title:              string;
    rating:             string;
    user_ratings_total: number;
    category:           string;
    icon:               string;
}